+++ CHANGELOG.md
@@ -11,6 +11,14 @@ auto-release workflow to deduplicate commits across successive runs.
 
 ## [Unreleased]
 
+## [5.158.3] - 2026-08-28
+
+### Fixed
+- block direct launchctl/systemctl service registration (#1311) (09de4da0)
+
+### Changed
+- migrate errorBox to errorCard, remove old component (#1337) (db0bb911)
+
 ## [5.158.2] - 2026-08-28
 
 ### Fixed